requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssutils==2.10.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.3
django-bower==5.2.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-decouple==2.1
django-dramatiq==0.11.6
django-extensions==3.2.3
django-filter==24.2
django-import-export==3.3.7
django-inlinecss==0.3.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-qrcode==0.3
django-redis==5.4.0
django-rest-framework==0.1.0
django-simple-history==3.5.0
django-timezone-field==6.1.0
django-webcam==0.4.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-datatables==0.7.1
dramatiq==1.16.0
et-xmlfile==1.1.0
frozenlist==1.4.1
future==1.0.0
geoip2==4.8.0
icalendar==5.0.12
idna==3.6
kombu==5.3.6
MarkupPy==1.14
maxminddb==2.6.0
multidict==6.0.5
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.1
pika==1.3.2
Pillow==10.3.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
pynliner==0.8.0
python-barcode==0.15.1
python-crontab==3.0.0
python-dateutil==2.9.0
pytz==2021.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.6.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
